refactor: src/utils.js → src/utils/index.js
refactor: move utils tests
